+++ include/ramdisk.h
@@ -13,6 +13,9 @@
 
 #include "device.h"
 
+/**
+ * @brief structure for manage the ramdisk
+ */
 struct ramdisk {
 	size_t size;
 	char *buffer;
@@ -22,7 +25,7 @@ struct ramdisk {
 int ramdisk_open(struct device *);
 ssize_t ramdisk_write(struct device *, struct device_request *);
 ssize_t ramdisk_read(struct device *, struct device_request *);
-ssize_t ramdisk_erase(struct device *, struct device_request *);
+int ramdisk_erase(struct device *, struct device_request *);
 int ramdisk_close(struct device *);
 
 int ramdisk_device_init(struct device *, uint64_t flags);
